bigfile/py: It is ok to have .fileh_open() as BigFile method
There was an XXX of whether fileh_open should be a BigFile method or
global function. However if it would be a global function it will need
to anyway accept file parameter to indicate which file is opened, and
that in turn suggests that it should be a file method. Remove XXX.