bigfile/py: Move PyVMA's support for cyclic GC close to pyvma_dealloc
The logic in pyvma_traverse and pyvma_clear needs to be synchronized
with PyVMA deallocation. In the next patche we'll be amending this
logic, and it will help a reader to keep all those functions together.

For the reference: PyVMA support for cyclic GC was introduced in
450ad804 (bigarray: ArrayRef support for BigArray). See also d97641d2
(bigfile/py: Properly untrack PyVMA from GC before dealloc).

/proposed-for-review-on https://lab.nexedi.com/nexedi/wendelin.core/merge_requests/12