zstor_2zurl:Fix ipv6 host for NEO/ZEO + more tests
This patch allows using WCFS with a NEO or ZEO storage which is
reachable by a URL which contains an ipv6 host.

Before this patch it was only possible to use WCFS with a NEO or ZEO
storage which is reachable by a URL which contains an ipv4 host.

This patch also adds one new test and extends another test. The tests
verify that 'zstor_2zurl' returns the expected URL.